pyproject.toml
[tool.poetry]
name = "mediarinhv2"
version = "0.1.0"
description = ""
authors = ["SooTher99 <tolikberdyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
email-validator = "^1.3.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.9"}
asyncpg = "^0.27.0"
alembic = "^1.10.3"
pillow = "^9.5.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
